+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## [Unreleased]
+
+## [Release (1.11.6)](https://github.com/mercadolibre/meli-card-drawer-ios/releases/tag/1.11.6)
+### Fixed
+- Fixed the card view's renderezation when has custom gradient.
+
+## [Release (1.11.5)](https://github.com/mercadolibre/meli-card-drawer-ios/releases/tag/1.11.5)
+## Fix
+- Fixed tag bottom corner radius
+
+## [Release (1.11.3)](https://github.com/mercadolibre/meli-card-drawer-ios/releases/tag/1.11.3)
+## Fix
+- Remove the available balance label when this info is not sent
+
 ## [Release (1.11.2)](https://github.com/mercadolibre/meli-card-drawer-ios/releases/tag/1.11.2)
 ## fix
 - changed combo switch redesign